Extract types related to routing
